Switch to 32 bit message indices when possible in aos
Larger queues need larger indices.  Everything is 32 bits to support old
32 bit CPUs.  Conditionally use 64 bit atomics when we can on 64 bit
processors so we can have more messages.  32 bit processors are becoming
pretty old, so no sense hamstringing the rest of the system based on
their constraints.

This also runs both tests on 64 bit processors to maintain coverage of
the 32 bit code.

+++ aos/configuration.h
@@ -209,9 +209,9 @@ bool ApplicationShouldStart(const Configuration *config, const Node *my_node,
                             const Application *application);
 
 // Returns the number of messages in the queue.
-int QueueSize(const Configuration *config, const Channel *channel);
-int QueueSize(size_t frequency,
-              std::chrono::nanoseconds channel_storage_duration);
+size_t QueueSize(const Configuration *config, const Channel *channel);
+size_t QueueSize(size_t frequency,
+                 std::chrono::nanoseconds channel_storage_duration);
 
 // Returns the number of scratch buffers in the queue.
 int QueueScratchBufferSize(const Channel *channel);
